my_streaming_project/dagster_producer/requirements.txt
dagster
dagit
kafka-python